:poop: Read three ATL11 pair tracks into datatree using for-loop
Not the most elegant solution, but manually reading the ATL11 pair track groups one by one in the for loop works. If the file pointer bug is fixed, a dictionary comprehension could be used to remove the for loop. Even better if the ValueError about malformed variables disappears, which would make this a one-liner!